+++ src/Factory/ObjectManagerFactory.php
@@ -10,8 +10,8 @@ use Coffreo\CephOdm\Entity\File;
 use Coffreo\CephOdm\EventListener\AddObjectListenerListener;
 use Coffreo\CephOdm\Persister\CephBucketPersister;
 use Coffreo\CephOdm\Persister\CephFilePersister;
-use Coffreo\CephOdm\Repository\CephBucketDataRepository;
-use Coffreo\CephOdm\Repository\CephFileDataRepository;
+use Coffreo\CephOdm\DataRepository\CephBucketDataRepository;
+use Coffreo\CephOdm\DataRepository\CephFileDataRepository;
 use Doctrine\Common\EventManager;
 use Doctrine\SkeletonMapper\Events;
 use Doctrine\SkeletonMapper\Hydrator\BasicObjectHydrator;
